+++ models/DataObject/AbstractObject.php
@@ -704,35 +704,37 @@ abstract class AbstractObject extends Model\Element\AbstractElement
     {
         $this->updateModificationInfos();
 
-        // save properties
-        $this->getProperties();
-        $this->getDao()->deleteAllProperties();
+        if ($this->isFieldDirty('properties')) {
+            // save properties
+            $properties = $this->getProperties();
+            $this->getDao()->deleteAllProperties();
 
-        foreach ($this->getProperties() as $property) {
-            if (!$property->getInherited()) {
-                $property->setDao(null);
-                $property->setCid($this->getId());
-                $property->setCtype('object');
-                $property->setCpath($this->getRealFullPath());
-                $property->save();
-            }
-        }
-
-        // save dependencies
-        $d = new Model\Dependency();
-        $d->setSourceType('object');
-        $d->setSourceId($this->getId());
-
-        foreach ($this->resolveDependencies() as $requirement) {
-            if ($requirement['id'] == $this->getId() && $requirement['type'] === 'object') {
-                // dont't add a reference to yourself
-                continue;
+            foreach ($properties as $property) {
+                if (!$property->getInherited()) {
+                    $property->setDao(null);
+                    $property->setCid($this->getId());
+                    $property->setCtype('object');
+                    $property->setCpath($this->getRealFullPath());
+                    $property->save();
+                }
             }
 
-            $d->addRequirement($requirement['id'], $requirement['type']);
-        }
+            // save dependencies
+            $d = new Model\Dependency();
+            $d->setSourceType('object');
+            $d->setSourceId($this->getId());
 
-        $d->save();
+            foreach ($this->resolveDependencies() as $requirement) {
+                if ($requirement['id'] == $this->getId() && $requirement['type'] === 'object') {
+                    // don't add a reference to yourself
+                    continue;
+                }
+
+                $d->addRequirement($requirement['id'], $requirement['type']);
+            }
+
+            $d->save();
+        }
 
         //set object to registry
         RuntimeCache::set(self::getCacheKey($this->getId()), $this);
